backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
openai==1.11.0
Pillow==10.2.0
requests==2.31.0
python-multipart==0.0.6
typing-extensions>=4.8.0
pydantic>=2.0.0
starlette>=0.35.0
